+++ src/services/axios/putProposal/putProposal.tsx
@@ -4,10 +4,15 @@ import {
   DEFAULT_PI,
   GENERAL,
   Projects,
+  RA_TYPE_EQUATORIAL,
   SKA_PHT_API_URL,
-  USE_LOCAL_DATA
+  USE_LOCAL_DATA,
+  VEL_TYPES
 } from '../../../utils/constants';
 import Proposal, { ProposalBackend } from '../../../utils/types/proposal';
+import { helpers } from '../../../utils/helpers';
+import Target, { TargetBackend } from 'utils/types/target';
+import { DocumentBackend, DocumentPDF } from 'utils/types/document';
 
 /*
 TODO:
@@ -44,7 +49,7 @@ function mappingPutProposal(proposal: Proposal, status: string) {
 
   const convertCategoryFormat = (_inValue: string): string => {
     const words = _inValue.split(' ');
-    const lowerCaseWords = words.map(word => word.charAt(0).toLowerCase() + word.slice(1));
+    const lowerCaseWords = words.map(word => word?.charAt(0)?.toLowerCase() + word.slice(1));
     const formattedString = lowerCaseWords.join('_');
     return formattedString;
   };
@@ -62,20 +67,77 @@ function mappingPutProposal(proposal: Proposal, status: string) {
     return subTypes;
   };
 
+  const getTargets = (targets: Target[]): TargetBackend[] => {
+    const outTargets = [];
+    for (let i = 0; i < targets.length; i++) {
+      const tar = targets[i];
+      const singlePointParam = tar.pointingPattern.parameters.find(
+        param => param.kind === 'SinglePointParameters'
+      );
+      const outTarget: TargetBackend = {
+        target_id: tar.name,
+        pointing_pattern: {
+          active: tar?.pointingPattern?.active,
+          parameters: [
+            {
+              kind: singlePointParam.kind,
+              offset_x_arcsec: singlePointParam.offsetXArcsec,
+              offset_y_arcsec: singlePointParam.offsetYArcsec
+            }
+          ]
+        },
+        reference_coordinate: {
+          kind: tar.referenceFrame === RA_TYPE_EQUATORIAL ? 'equatorial' : 'galactic',
+          ra: Number(tar.ra),
+          dec: Number(tar.dec),
+          unit: [tar.raUnit, tar.decUnit],
+          reference_frame: tar.rcReferenceFrame
+        },
+        radial_velocity: {
+          quantity: {
+            value: Number(tar.vel),
+            unit: tar.velUnit
+          },
+          definition: VEL_TYPES.find(item => item.value === tar.velType).label,
+          reference_frame: tar.raReferenceFrame,
+          redshift: Number(tar.redshift)
+        }
+      };
+      outTargets.push(outTarget);
+    }
+    return outTargets;
+  };
+
+  const getDocuments = (sciencePDF: DocumentPDF, technicalPDF: DocumentPDF): DocumentBackend[] => {
+    const documents = [];
+    if (sciencePDF?.link) {
+      documents.push({
+        document_id: sciencePDF.documentId,
+        link: sciencePDF?.link,
+        type: 'proposal_science'
+      });
+    }
+    if (technicalPDF?.link) {
+      documents.push({
+        document_id: technicalPDF?.documentId,
+        link: technicalPDF?.link,
+        type: 'proposal_technical'
+      });
+    }
+    return documents;
+  };
+
   // TODO : complete mapping for all properties
   const transformedProposal: ProposalBackend = {
-    /*
-      science_category: GENERAL.ScienceCategory?.find(
-        category => category.value === proposal?.category
-      )?.label,
-      */
-    prsl_id: proposal?.id?.toString(),
+    prsl_id: proposal?.id,
     status: status,
-    submitted_on: '', // TODO
-    submitted_by: '', // TODO
-    investigator_refs: [DEFAULT_PI.id],
+    submitted_on: '', // TODO // to fill for submit
+    submitted_by: '', // TODO // to fill for submit
+    investigator_refs: proposal.team?.map(investigator => {
+      return investigator.id;
+    }),
     metadata: {
-      version: proposal.version++,
+      version: proposal.version + 1,
       created_by: proposal.createdBy,
       created_on: proposal.createdOn,
       last_modified_by: `${DEFAULT_PI.firstName} ${DEFAULT_PI.lastName}`,
@@ -88,45 +150,31 @@ function mappingPutProposal(proposal: Proposal, status: string) {
         main_type: convertCategoryFormat(Projects.find(p => p.id === proposal.proposalType).title),
         sub_type: getSubCategory(proposal.proposalType, proposal.proposalSubType)
       },
-      abstract: '',
-      science_category: '',
-      targets: [],
-      documents: [],
-      investigators: [
-        {
-          investigator_id: DEFAULT_PI.id,
-          given_name: DEFAULT_PI.firstName,
-          family_name: DEFAULT_PI.lastName,
-          email: DEFAULT_PI.email,
-          organization: DEFAULT_PI.affiliation,
-          for_phd: DEFAULT_PI.phdThesis,
-          principal_investigator: DEFAULT_PI.pi
-        }
-      ],
+      abstract: proposal.abstract,
+      science_category: GENERAL.ScienceCategory?.find(
+        category => category.value === proposal?.scienceCategory
+      )?.label,
+      targets: getTargets(proposal.targets),
+      documents: getDocuments(proposal.sciencePDF, proposal.technicalPDF), // TODO check file upload issue
+      investigators: proposal.team.map(teamMember => {
+        return {
+          investigator_id: teamMember.id,
+          given_name: teamMember.firstName,
+          family_name: teamMember.lastName,
+          email: teamMember.email,
+          organization: teamMember.affiliation,
+          for_phd: teamMember.phdThesis,
+          principal_investigator: teamMember.pi
+        };
+      }),
       observation_sets: [], // TODO add a conversion function to change units to 'm/s' when mapping so we don't have a 'm / s' format in front-end
       data_product_sdps: [],
       data_product_src_nets: [],
       results: []
-
-      /*
-      targets: proposal?.targets?.map(target => ({
-        name: target?.name,
-        right_ascension: target?.ra,
-        declination: target?.dec,
-        velocity: parseFloat(target?.vel),
-        velocity_unit: '', // TODO: confirm what units should be expected
-        right_ascension_unit: '', // TODO: confirm what units should be expected
-        declination_unit: '' // TODO: confirm what units should be expected
-      })),
-      */
-      /*
-      investigators: proposal.team?.map(teamMember => ({
-      science_programmes: scienceProgrammes
-      */
     }
   };
   // trim undefined properties
-  this.trimObject(transformedProposal);
+  helpers.transform.trimObject(transformedProposal);
   return transformedProposal;
 }
 
@@ -144,7 +192,7 @@ async function PutProposal(proposal, status?) {
       convertedProposal,
       AXIOS_CONFIG
     );
-    return typeof result === 'undefined' ? 'error.API_UNKNOWN_ERROR' : result.data;
+    return typeof result === 'undefined' ? 'error.API_UNKNOWN_ERROR' : result; // result?.data;
   } catch (e) {
     return { error: e.message };
   }