+++ src/utils/types/teamMember.tsx
@@ -12,7 +12,7 @@ export type TeamMemberBackend = {
 */
 
 type TeamMember = {
-  id: number;
+  id: string;
   firstName: string;
   lastName: string;
   email: string;