+++ CHANGELOG.md
@@ -6,7 +6,11 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 Future
 
 ---
-
+* STAR-615: Proposal mapping:
+   - PutProposal mapping (partial - observation_sets, data_product_sdps ,data_product_src_nets and results still to map)
+   - Update SensCalc display results to display sensitivity or integration time
+   - Update pages accordingly for breaking changes
+   - Update proposal frontend types
 * STAR-606: Proposal mapping:
    - GetProposal mapping
    - Update proposal frontend types